Allow API client class to receive Guzzle config
API Client class is basically Guzzle Client's wrapper but it doesn't
have a hook to pass default configuration to the Guzzle Client.

This adds an argument to API Client constructor which allows to pass
Guzzle Client configuration to Guzzle Client.

This is useful to avoid curl's unrecognized content encoding type error
by passing decode_content false configuration to the client.